refactor comment api
