+++ interfaces/LikeDao.ts
@@ -1,6 +1,6 @@
 import Tuit from "../models/Tuit";
 import Like from "../models/Like";
-import User from "../models/Like";
+import User from "../models/User";
 
 /**
  * @file This file is an interface for using likes collections
